Make 'MultiResHashTable' 'nn.Module'
