+++ torch_nerf/configs/network/instant_nerf.yaml
@@ -1,2 +1,8 @@
+type: instant_nerf
 pos_dim: 3
 view_dir_dim: 3
+num_level: 16
+log_max_entry_per_level: 19
+table_feat_dim: 2
+min_res: 16
+max_res: 4096  # this number is arbitrary
