+++ torch_nerf/configs/network/nerf.yaml
@@ -1,2 +1,3 @@
+type: nerf
 pos_dim: 3
 view_dir_dim: 3
