+++ torch_nerf/src/scene/primitives/__init__.py
@@ -1,3 +1,2 @@
 from torch_nerf.src.scene.primitives.primitive_base import PrimitiveBase
 from torch_nerf.src.scene.primitives.cube import PrimitiveCube
-from torch_nerf.src.scene.primitives.hash_table import PrimitiveHashEncoding
